+++ test/runtests.jl
@@ -19,6 +19,7 @@ using PolySeries
 
     @testset "Multiplication Reference" begin
         include("mul_reference_tests.jl")
+        include("square_kernel_tests.jl")
     end
 
     @testset "Type Stability" begin